+++ editors/tree-sitter-wirescript/queries/highlights.scm
@@ -82,6 +82,8 @@
 
 "unsafe" @keyword.modifier
 
+"is" @keyword.operator
+
 ; ------------------------------------------------------------------- literals
 
 (integer) @number